pyproject.toml
[tool.poetry]
name = "module-6-apop01100"
version = "0.1.0"
description = ""
authors = ["thoriqafif <mthoriqafifr01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
sqlalchemy-serializer = "^1.4.22"
flasgger = "^0.9.7.1"
pytest-cov = "^5.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
